Fix premature GC issues in several constructors.
Several constructors would store pointers to newly allocated ruby objects inside an unwrapped C struct on the heap.
This, coupled with optimizing compilers, made it possible that these objects were not visible to the GC in the mark phase and prematurely collected, resulting in dangling pointers, crashes due to segfaults, etc.

These changes aim to make sure that the objects will be always visible to the GC, by storing pointers to them in local variables protected by RB_GC_GUARD.